+++ src/options.ts
@@ -1,10 +1,11 @@
+import type { ReconnectPolicy } from './room/ReconnectPolicy';
 import type {
   AudioCaptureOptions,
+  AudioOutputOptions,
   TrackPublishDefaults,
   VideoCaptureOptions,
 } from './room/track/options';
 import type { AdaptiveStreamSettings } from './room/track/types';
-import type { ReconnectPolicy } from './room/ReconnectPolicy';
 
 /**
  * @internal
@@ -44,6 +45,11 @@ export interface InternalRoomOptions {
   publishDefaults?: TrackPublishDefaults;
 
   /**
+   * audio output for the room
+   */
+  audioOutput?: AudioOutputOptions;
+
+  /**
    * should local tracks be stopped when they are unpublished. defaults to true
    * set this to false if you would prefer to clean up unpublished local tracks manually.
    */
@@ -80,6 +86,9 @@ export interface RoomOptions extends Partial<InternalRoomOptions> {}
 export interface InternalRoomConnectOptions {
   /** autosubscribe to room tracks after joining, defaults to true */
   autoSubscribe: boolean;
+
+  /** amount of time for PeerConnection to be established, defaults to 15s */
+  peerConnectionTimeout: number;
 
   /**
    * use to override any RTCConfiguration options.