+++ src/room/defaults.ts
@@ -40,4 +40,5 @@ export const roomOptionDefaults: InternalRoomOptions = {
 export const roomConnectOptionDefaults: InternalRoomConnectOptions = {
   autoSubscribe: true,
   maxRetries: 1,
+  peerConnectionTimeout: 15_000,
 } as const;