+++ src/room/track/RemoteAudioTrack.ts
@@ -1,7 +1,10 @@
+import log from '../../logger';
+import { TrackEvent } from '../events';
 import { AudioReceiverStats, computeBitrate } from '../stats';
+import { supportsSetSinkId } from '../utils';
+import type { AudioOutputOptions } from './options';
 import RemoteTrack from './RemoteTrack';
 import { Track } from './Track';
-import log from '../../logger';
 
 export default class RemoteAudioTrack extends RemoteTrack {
   private prevStats?: AudioReceiverStats;
@@ -16,15 +19,21 @@ export default class RemoteAudioTrack extends RemoteTrack {
 
   private webAudioPluginNodes: AudioNode[];
 
+  private sinkId?: string;
+
   constructor(
     mediaTrack: MediaStreamTrack,
     sid: string,
     receiver?: RTCRtpReceiver,
     audioContext?: AudioContext,
+    audioOutput?: AudioOutputOptions,
   ) {
     super(mediaTrack, sid, Track.Kind.Audio, receiver);
     this.audioContext = audioContext;
     this.webAudioPluginNodes = [];
+    if (audioOutput) {
+      this.sinkId = audioOutput.deviceId;
+    }
   }
 
   /**
@@ -57,6 +66,23 @@ export default class RemoteAudioTrack extends RemoteTrack {
     return highestVolume;
   }
 
+  /**
+   * calls setSinkId on all attached elements, if supported
+   * @param deviceId audio output device
+   */
+  async setSinkId(deviceId: string) {
+    this.sinkId = deviceId;
+    await Promise.all(
+      this.attachedElements.map((elm) => {
+        if (!supportsSetSinkId(elm)) {
+          return;
+        }
+        /* @ts-ignore */
+        return elm.setSinkId(deviceId) as Promise<void>;
+      }),
+    );
+  }
+
   attach(): HTMLMediaElement;
   attach(element: HTMLMediaElement): HTMLMediaElement;
   attach(element?: HTMLMediaElement): HTMLMediaElement {
@@ -68,6 +94,10 @@ export default class RemoteAudioTrack extends RemoteTrack {
     }
     if (this.elementVolume) {
       element.volume = this.elementVolume;
+    }
+    if (this.sinkId && supportsSetSinkId(element)) {
+      /* @ts-ignore */
+      element.setSinkId(this.sinkId);
     }
     if (this.audioContext && needsNewWebAudioConnection) {
       log.debug('using audio context mapping');
@@ -96,9 +126,12 @@ export default class RemoteAudioTrack extends RemoteTrack {
     } else {
       detached = super.detach(element);
       // if there are still any attached elements after detaching, connect webaudio to the first element that's left
-      if (this.audioContext && this.attachedElements.length > 0) {
+      // disconnect webaudio otherwise
+      if (this.audioContext) {
         if (this.attachedElements.length > 0) {
           this.connectWebAudio(this.audioContext, this.attachedElements[0]);
+        } else {
+          this.disconnectWebAudio();
         }
       }
     }
@@ -109,10 +142,12 @@ export default class RemoteAudioTrack extends RemoteTrack {
    * @internal
    * @experimental
    */
-  setAudioContext(audioContext: AudioContext) {
+  setAudioContext(audioContext: AudioContext | undefined) {
     this.audioContext = audioContext;
-    if (this.attachedElements.length > 0) {
+    if (audioContext && this.attachedElements.length > 0) {
       this.connectWebAudio(audioContext, this.attachedElements[0]);
+    } else if (!audioContext) {
+      this.disconnectWebAudio();
     }
   }
 
@@ -143,6 +178,23 @@ export default class RemoteAudioTrack extends RemoteTrack {
 
     if (this.elementVolume) {
       this.gainNode.gain.setTargetAtTime(this.elementVolume, 0, 0.1);
+    }
+
+    // try to resume the context if it isn't running already
+    if (context.state !== 'running') {
+      context
+        .resume()
+        .then(() => {
+          if (context.state !== 'running') {
+            this.emit(
+              TrackEvent.AudioPlaybackFailed,
+              new Error("Audio Context couldn't be started automatically"),
+            );
+          }
+        })
+        .catch((e) => {
+          this.emit(TrackEvent.AudioPlaybackFailed, e);
+        });
     }
   }